refactor(hw/display): moved to hw/display/*
